Refactor logic for getting samples + add test
