+++ src/encoded/item_utils/utils.py
@@ -3,8 +3,9 @@ from typing import Any, Callable, Dict, List, Optional
 
 from dcicutils import ff_utils
 from pyramid.request import Request
+from webtest import TestApp
 
-from ..utils import get_item as get_item_from_request
+from ..utils import get_item as get_item_from_request, get_item_from_testapp
 
 
 @dataclass(frozen=True)
@@ -16,10 +17,11 @@ class RequestHandler:
 
     request: Optional[Request] = None
     auth_key: Optional[Dict[str, str]] = None
+    test_app: Optional[TestApp] = None
 
     def __post_init__(self) -> None:
-        if not self.request and not self.auth_key:
-            raise ValueError("Either request or auth_key must be provided")
+        if not self.request and not self.auth_key and not self.test_app:
+            raise ValueError("Either request, auth_key, or test app must be provided")
 
     def get_items(
         self, identifiers: List[str], collection: Optional[str] = None
@@ -31,10 +33,16 @@ class RequestHandler:
                 self._get_item_from_request(identifier, collection=collection)
                 for identifier in unique_identifiers
             ]
-        return [
-            self._get_item_from_auth_key(identifier)
-            for identifier in unique_identifiers
-        ]
+        if self.auth_key:
+            return [
+                self._get_item_from_auth_key(identifier)
+                for identifier in unique_identifiers
+            ]
+        if self.test_app:
+            return [
+                self._get_item_from_test_app(identifier, collection=collection)
+                for identifier in unique_identifiers
+            ]
 
     def get_item(
         self, identifier: str, collection: Optional[str] = None
@@ -44,7 +52,10 @@ class RequestHandler:
             return {}
         if self.request:
             return self._get_item_from_request(identifier, collection=collection)
-        return self._get_item_from_auth_key(identifier)
+        if self.auth_key:
+            return self._get_item_from_auth_key(identifier)
+        if self.test_app:
+            return self._get_item_from_test_app(identifier, collection=collection)
 
     def _get_item_from_request(
         self, identifier: str, collection: Optional[str] = None
@@ -56,6 +67,14 @@ class RequestHandler:
         """Get item from auth_key"""
         return ff_utils.get_metadata(
             identifier, key=self.auth_key, add_on="frame=object"
+        )
+
+    def _get_item_from_test_app(
+        self, identifier: str, collection: Optional[str] = None
+    ) -> Dict[str, Any]:
+        """Get item from test app"""
+        return get_item_from_testapp(
+            self.test_app, identifier, collection=collection, frame="object"
         )
